fix: use reverse proxy as seperate service :wrench:
The previous Dockerfile configuration for the server was using the nginx reverse proxy as a build stage which was not working because the ENTRYPOINT for the server was interfering with the ENTRYPOINT for nginx. Also the 'copy --from=nginx:1.18-alpine' was pulling the nginx image but allowing it to be available for subsequent commands with it. This removes using reverse proxy as a stage in server builds and configures the server reverse proxy as a seperate service.